moves BaseSelect to Sidebar level folder
