+++ packages/vizbuilder/src/components/Sidebar/ConditionManager/ConditionPropertySelect.jsx
@@ -3,7 +3,7 @@ import classnames from "classnames";
 import escapeRegExp from "lodash/escapeRegExp";
 import {Icon} from "@blueprintjs/core";
 
-import BaseSelect from "../../BaseSelect";
+import BaseSelect from "../BaseSelect";
 import {composePropertyName} from "../../../helpers/formatting";
 
 function ConditionPropertySelect(props) {
